AGRISYNC/requirements.txt
# Core dependencies
streamlit>=1.22.0
numpy>=1.22.0
pandas>=1.4.0
pillow>=9.0.0
plotly>=5.6.0
pydub>=0.25.1
requests>=2.27.1
scipy>=1.8.0

# Machine learning
scikit-learn>=1.1.0
tensorflow>=2.9.0
torch>=1.11.0
torchvision>=0.12.0
transformers>=4.18.0

# Audio processing
librosa>=0.9.2
SpeechRecognition>=3.8.1
gTTS>=2.2.4

# Image processing
opencv-python-headless>=4.5.5
efficientnet>=1.0.0

# Data visualization
matplotlib>=3.5.1
seaborn>=0.11.2

# Weather data
meteomatics>=1.0.1